more styling, refactored components (minor)
