+++ public/index.html
@@ -15,7 +15,7 @@
   <meta name="viewport" content="width=device-width, initial-scale=1" />
   <meta name="theme-color" content="#000000" />
   <meta name="description" content="Find a bike share station in a city of your choice" />
-  <title>Bike Share Finer</title>
+  <title>Bike Share Finder</title>
 </head>
 
 <body>
