+++ src/BikeSVG.js
@@ -5,12 +5,15 @@ const BikeSVG = (props) => (
         y={0}
         x={0}
         xmlns="http://www.w3.org/2000/svg"
+        id="\u5716\u5C64_1"
         style={{
             height: "100%",
             width: "100%",
-            background: "#f1f2f3",
+            background: "#f1f2f3 0 0",
             shapeRendering: "auto",
         }}
+        width={200}
+        height={200}
         {...props}
     >
         <g
@@ -25,8 +28,10 @@ const BikeSVG = (props) => (
                     <g className="ldl-ani">
                         <g
                             style={{
-                                transform: "scale(.91)",
                                 transformOrigin: "50px 50px 0",
+                                transform: "translate(-1px)",
+                                animation:
+                                    "1.11111s linear -.625s infinite normal forwards running wander-h-f87635cd-e757-4a71-9327-b5839fbfee6f",
                             }}
                             className="ldl-layer"
                         >
@@ -39,17 +44,18 @@ const BikeSVG = (props) => (
                                 cy={61.7}
                                 cx={21.1}
                                 style={{
-                                    fill: "#fff",
-                                    stroke: "#333",
+                                    fill: "#264653",
+                                    stroke: "#e76f51",
                                 }}
                                 className="ldl-ani"
                             />
                         </g>
                         <g
                             style={{
-                                transform: "scale(.91)",
                                 transformOrigin: "50px 50px 0",
-
+                                transform: "translate(-1px)",
+                                animation:
+                                    "1.11111s linear -.694444s infinite normal forwards running wander-h-f87635cd-e757-4a71-9327-b5839fbfee6f",
                             }}
                             className="ldl-layer"
                         >
@@ -59,7 +65,7 @@ const BikeSVG = (props) => (
                                 cy={61.7}
                                 cx={21.1}
                                 style={{
-                                    fill: "#e0e0e0",
+                                    fill: "#2a9d8f",
                                 }}
                                 className="ldl-ani"
                             />
@@ -70,9 +76,10 @@ const BikeSVG = (props) => (
                     <g className="ldl-ani">
                         <g
                             style={{
-                                transform: "scale(.91)",
                                 transformOrigin: "50px 50px 0",
-
+                                transform: "translate(-1px)",
+                                animation:
+                                    "1.11111s linear -.763889s infinite normal forwards running wander-h-f87635cd-e757-4a71-9327-b5839fbfee6f",
                             }}
                             className="ldl-layer"
                         >
@@ -85,17 +92,18 @@ const BikeSVG = (props) => (
                                 cy={61.7}
                                 cx={78.9}
                                 style={{
-                                    fill: "#fff",
-                                    stroke: "#333",
+                                    fill: "#264653",
+                                    stroke: "#e76f51",
                                 }}
                                 className="ldl-ani"
                             />
                         </g>
                         <g
                             style={{
-                                transform: "scale(.91)",
                                 transformOrigin: "50px 50px 0",
-
+                                transform: "translate(-1px)",
+                                animation:
+                                    "1.11111s linear -.833333s infinite normal forwards running wander-h-f87635cd-e757-4a71-9327-b5839fbfee6f",
                             }}
                             className="ldl-layer"
                         >
@@ -105,7 +113,7 @@ const BikeSVG = (props) => (
                                 cy={61.7}
                                 cx={78.9}
                                 style={{
-                                    fill: "#e0e0e0",
+                                    fill: "#2a9d8f",
                                 }}
                                 className="ldl-ani"
                             />
@@ -116,9 +124,10 @@ const BikeSVG = (props) => (
                     <g className="ldl-ani">
                         <g
                             style={{
-                                transform: "scale(.91)",
                                 transformOrigin: "50px 50px 0",
-
+                                transform: "translate(-1px)",
+                                animation:
+                                    "1.11111s linear -.902778s infinite normal forwards running wander-h-f87635cd-e757-4a71-9327-b5839fbfee6f",
                             }}
                             className="ldl-layer"
                         >
@@ -126,16 +135,17 @@ const BikeSVG = (props) => (
                                 fill="#77a4bd"
                                 d="M80.3 61 63.7 24.1c-.3-.8-1.3-1.1-2-.8-.8.3-1.1 1.3-.8 2l6.3 14H37.5L35 33.7 32.2 35l2.8 6.2-14.2 18.4c-.5.7-.6 1.6-.2 2.3.4.8 1.1 1.2 2 1.2h23.3l1-.4 22.3-19.1 8.3 18.5c.3.6.8.9 1.4.9.2 0 .4 0 .6-.1.8-.2 1.2-1.1.8-1.9zM36.4 44.4l7.1 15.8H24.3l12.1-15.8zm10 14.8-7.5-16.7h27L46.4 59.2z"
                                 style={{
-                                    fill: "#77a4bd",
+                                    fill: "#2a9d8f",
                                 }}
                                 className="ldl-ani"
                             />
                         </g>
                         <g
                             style={{
-                                transform: "scale(.91)",
                                 transformOrigin: "50px 50px 0",
-
+                                transform: "translate(-1px)",
+                                animation:
+                                    "1.11111s linear -.972222s infinite normal forwards running wander-h-f87635cd-e757-4a71-9327-b5839fbfee6f",
                             }}
                             className="ldl-layer"
                         >
@@ -148,15 +158,17 @@ const BikeSVG = (props) => (
                                 stroke="#c33737"
                                 fill="none"
                                 style={{
-                                    stroke: "#c33737",
+                                    stroke: "#e76f51",
                                 }}
                                 className="ldl-ani"
                             />
                         </g>
                         <g
                             style={{
-                                transform: "scale(.91)",
                                 transformOrigin: "50px 50px 0",
+                                transform: "translate(-1px)",
+                                animation:
+                                    "1.11111s linear -1.04167s infinite normal forwards running wander-h-f87635cd-e757-4a71-9327-b5839fbfee6f",
                             }}
                             className="ldl-layer"
                         >
@@ -169,16 +181,18 @@ const BikeSVG = (props) => (
                                 cy={61}
                                 cx={45.5}
                                 style={{
-                                    fill: "#a0c8d7",
-                                    stroke: "#77a4bd",
+                                    fill: "#e9c46a",
+                                    stroke: "#2a9d8f",
                                 }}
                                 className="ldl-ani"
                             />
                         </g>
                         <g
                             style={{
-                                transform: "scale(.91)",
                                 transformOrigin: "50px 50px 0",
+                                transform: "translate(-1px)",
+                                animation:
+                                    "1.11111s linear -1.11111s infinite normal forwards running wander-h-f87635cd-e757-4a71-9327-b5839fbfee6f",
                             }}
                             className="ldl-layer"
                         >
@@ -191,7 +205,7 @@ const BikeSVG = (props) => (
                                 stroke="#c33737"
                                 fill="none"
                                 style={{
-                                    stroke: "#c33737",
+                                    stroke: "#e76f51",
                                 }}
                                 className="ldl-ani"
                             />
@@ -200,6 +214,11 @@ const BikeSVG = (props) => (
                 </g>
             </g>
         </g>
+        <style id="wander-h-f87635cd-e757-4a71-9327-b5839fbfee6f">
+            {
+                "@keyframes wander-h-f87635cd-e757-4a71-9327-b5839fbfee6f{0%{animation-timing-function:cubic-bezier(.6599,.1803,.8238,.6733);transform:translate(-1px,0)}10%{animation-timing-function:cubic-bezier(.4126,.3294,.7102,.6652);transform:translate(-.795px,0)}17.5%{animation-timing-function:cubic-bezier(.3486,.3328,.6192,.6682);transform:translate(-.438px,0)}35.1%{animation-timing-function:cubic-bezier(.2708,.3048,.6411,.6958);transform:translate(.5760000000000001px,0)}40.1%{animation-timing-function:cubic-bezier(.1731,.3255,.3395,.8291);transform:translate(.799px,0)}50.1%{animation-timing-function:cubic-bezier(.6576,.1906,.8206,.6689);transform:translate(1px,0)}60.1%{animation-timing-function:cubic-bezier(.4092,.3294,.7101,.6653);transform:translate(.793px,0)}67.6%{animation-timing-function:cubic-bezier(.3528,.3327,.616,.6682);transform:translate(.437px,0)}85%{animation-timing-function:cubic-bezier(.2925,.3308,.6163,.6733);transform:translate(-.5740000000000001px,0)}90%{animation-timing-function:cubic-bezier(.1781,.33,.3368,.8179);transform:translate(-.796px,0)}to{transform:translate(-1px,0)}}"
+            }
+        </style>
     </svg>
 )
